Added options to make LAN access possible
The immich app has a feature to switch to another URL when on a specific
wifi network (e.g. home network). This change adds an option for listen
address so that the server can listen on LAN traffic. For this to work
with authelia, we need to add the appropriate redirect URL as well.

This is very handy for those of us who are using a proxy which blocks
large file uploads -- large files will then get synced automatically
when connecting to the home network.